+++ xlive/Blam/Engine/interface/user_interface_controller.cpp
@@ -576,7 +576,7 @@ static void user_interface_controller_boot_to_dash_check(void)
 			//_user_interface_channel_type_dialog, //orignally this , but changed to hardware_error so it can overrule disconnection screen
 			_user_interface_channel_type_hardware_error,
 			_window_4,
-			NONE,
+			(uint16)NONE,
 			exit_callback,
 			error_id);
